[sw] Warn on sign conversions
Recent versions of Clang default to warning on sign conversions
(`-Wsign-conversion`) when `-Wconversion` is enabled, which we do enable
in OpenTitan. We can either resolve those cases (by making the
conversion explicit) or disable the warning. This patch tries to resolve
the warnings.

+++ sw/device/lib/base/hardened_memory.c
@@ -153,7 +153,7 @@ hardened_bool_t consttime_memeq_byte(const void *lhs, const void *rhs,
 
     // Same as above. The compiler can cache the value of `a[offset]`, but it
     // has no chance to strength-reduce this operation.
-    ones = launder32(ones) & (launder32((uint32_t)*a) ^ ~*b);
+    ones = launder32(ones) & (launder32((uint32_t)*a) ^ ~(uint32_t)*b);
   }
 
   HARDENED_CHECK_EQ(count, len);
